fix: Apply MCP initialization fix to all MCP clients
Extends the MCP initialization handshake fix from Git MCP to all MCP clients
to prevent RPC error -32602 across the entire system.

**Changes:**
- Create shared BaseMCPClient with proper MCP initialization handshake
- Refactor GitMCPClient to extend BaseMCPClient (maintains fix from #75)
- Refactor FileSystemMCPClient to extend BaseMCPClient
- Refactor DockerMCPClient to extend BaseMCPClient
- Refactor TestingMCPClient to extend BaseMCPClient
- Eliminate code duplication across all MCP clients
- Add proper context manager support to all clients

**Technical Details:**
All MCP servers use the official MCP library requiring proper protocol initialization:
1. Send 'initialize' request with protocol version and capabilities
2. Send 'initialized' notification after successful response
3. Only then can tools be called without RPC errors

**Impact:**
- Prevents issue #75 from occurring on any MCP server
- Consistent MCP protocol compliance across all clients
- Reduced code duplication and easier maintenance
- All existing tests continue to pass
