+++ src/mcp_servers/docker/launcher.py
@@ -2,13 +2,12 @@
 
 import argparse
 import subprocess
-import sys
-import json
 from pathlib import Path
-from typing import Optional, Dict, Any
+from typing import Optional
+from ..base_client import BaseMCPClient
 
 
-class DockerMCPLauncher:
+class DockerMCPLauncher(BaseMCPClient):
     """Launcher for Docker MCP Server with stdio communication."""
 
     def __init__(self, project_root: Optional[str] = None):
@@ -18,37 +17,11 @@ class DockerMCPLauncher:
             project_root: Root directory for Docker operations.
                          If None, uses current working directory.
         """
-        self.project_root = project_root or str(Path.cwd())
-        self.process: Optional[subprocess.Popen] = None
-
-    def start_server(self) -> bool:
-        """Start the Docker MCP Server as a subprocess.
-
-        Returns:
-            True if server started successfully, False otherwise.
-        """
-        try:
-            server_module = Path(__file__).parent / "server.py"
-
-            # Start server process
-            cmd = [sys.executable, str(server_module)]
-            if self.project_root:
-                cmd.append(self.project_root)
-
-            self.process = subprocess.Popen(
-                cmd,
-                stdin=subprocess.PIPE,
-                stdout=subprocess.PIPE,
-                stderr=subprocess.PIPE,
-                text=True,
-                bufsize=0,  # Unbuffered for real-time communication
-            )
-
-            return True
-
-        except Exception as e:
-            print(f"Failed to start Docker MCP Server: {e}")
-            return False
+        super().__init__(project_root, "Docker MCP Server")
+    
+    def _get_server_script_path(self) -> Path:
+        """Get the path to the Docker server script."""
+        return Path(__file__).parent / "server.py"
 
     def start(self) -> subprocess.Popen:
         """Start the Docker MCP Server as a subprocess (legacy method).
@@ -60,74 +33,9 @@ class DockerMCPLauncher:
             return self.process
         raise RuntimeError("Failed to start Docker MCP Server")
 
-    def stop_server(self) -> None:
-        """Stop the Docker MCP Server process."""
-        if self.process:
-            self.process.terminate()
-            try:
-                self.process.wait(timeout=5)
-            except subprocess.TimeoutExpired:
-                self.process.kill()
-                self.process.wait()
-            finally:
-                self.process = None
-
     def stop(self) -> None:
         """Stop the Docker MCP Server process (legacy method)."""
         self.stop_server()
-
-    def send_request(
-        self, method: str, params: Optional[Dict[str, Any]] = None
-    ) -> Optional[Dict[str, Any]]:
-        """Send a JSON-RPC request to the server.
-
-        Args:
-            method: Method name to call
-            params: Parameters for the method
-
-        Returns:
-            Response from server or None if error
-        """
-        if not self.process:
-            return None
-
-        request = {"jsonrpc": "2.0", "id": 1, "method": method, "params": params or {}}
-
-        try:
-            # Send request
-            request_json = json.dumps(request) + "\n"
-            if self.process.stdin:
-                self.process.stdin.write(request_json)
-                self.process.stdin.flush()
-
-            # Read response
-            if self.process.stdout:
-                response_line = self.process.stdout.readline()
-                if response_line:
-                    return json.loads(response_line.strip())
-
-        except Exception as e:
-            print(f"Error communicating with Docker MCP server: {e}")
-
-        return None
-
-    def list_tools(self) -> Optional[Dict[str, Any]]:
-        """List available tools."""
-        return self.send_request("tools/list")
-
-    def call_tool(
-        self, name: str, arguments: Dict[str, Any]
-    ) -> Optional[Dict[str, Any]]:
-        """Call a specific tool.
-
-        Args:
-            name: Tool name
-            arguments: Tool arguments
-
-        Returns:
-            Tool result
-        """
-        return self.send_request("tools/call", {"name": name, "arguments": arguments})
 
     def is_running(self) -> bool:
         """Check if the Docker MCP Server process is running."""
